feat: add Chopin creation skill
