+++ skills/creating-chopin-plans/prompt.md
@@ -1,0 +1,11 @@
+# Create a Chopin plan
+
+The conversation below has settled enough context for one initial Chopin
+document. Follow the installed `creating-chopin-plans` skill and the current
+Chopin MCP initialize instructions and tool descriptions.
+
+<SETTLED_CONVERSATION>
+
+...
+
+</SETTLED_CONVERSATION>
